requirements.txt
click==8.1.3
Flask==2.2.3
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
psycopg2==2.9.5
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
Werkzeug==2.2.3
zipp==3.13.0
